backend/requirements.txt
black==23.9.1
Django==4.2.5
django-behaviors==0.5.1
djangorestframework==3.14.0
flake8==6.1.0
gunicorn==21.2.0
isort==5.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
